backend/api_auth/pyproject.toml
[tool.poetry]
name = "api-auth"
version = "0.1.0"
description = ""
authors = ["varg <vardeathyandex.ru>"]
readme = "README.md"
packages = [{include = "api_auth"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.85.1"
uvicorn = "^0.19.0"
databases = "^0.6.1"
aioredis = "^2.0.1"
dotenv = "^0.0.5"
python-dotenv = "^0.21.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
orjson = "^3.8.0"
python-multipart = "^0.0.5"
asyncpg = "^0.26.0"
psycopg2-binary = "^2.9.4"
fastapi-mail = {extras = ["httpx"], version = "^1.2.0"}
alembic = "^1.8.1"
sqlmodel = "^0.0.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
